Implement a complete CRUD on TreeNode
* move package org.seedstack.coffig.data -> org.Seedstack.coffig
* Improve property not found message
