+++ src/main/java/org/seedstack/coffig/MutableValueNode.java
@@ -5,7 +5,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package org.seedstack.coffig.data;
+package org.seedstack.coffig;
 
 public class MutableValueNode extends ValueNode implements MutableTreeNode {
     public MutableValueNode(String value) {
