+++ src/test/java/org/seedstack/coffig/MapNodeTest.java
@@ -5,12 +5,11 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package org.seedstack.coffig.data;
+package org.seedstack.coffig;
 
 import org.assertj.core.api.Assertions;
 import org.junit.Test;
-import org.seedstack.coffig.ConfigurationException;
-import org.seedstack.coffig.PropertyNotFoundException;
+import org.seedstack.coffig.*;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -27,9 +26,16 @@ public class MapNodeTest {
             mapNode1.value("name");
             Assertions.failBecauseExceptionWasNotThrown(PropertyNotFoundException.class);
         } catch (PropertyNotFoundException e) {
-            assertThat(e).hasMessage("Property \"name\" was not found");
-            assertThat(e.getPropertyName()).isEqualTo("name");
+            assertThat(e).hasMessage("Property \"[name]\" was not found");
+            assertThat(e.getPropertyName()).isEqualTo("[name]");
         }
+    }
+
+    @Test
+    public void testNodeExist() {
+        MapNode mapNode1 = new MapNode(new NamedNode("id", "foo"));
+        Assertions.assertThat(mapNode1.exist("id")).isTrue();
+        Assertions.assertThat(mapNode1.exist("zzz")).isFalse();
     }
 
     @Test(expected = UnsupportedOperationException.class)
@@ -91,17 +97,23 @@ public class MapNodeTest {
     }
 
     @Test
-    public void testSearch() throws Exception {
+    public void testGet() throws Exception {
         MapNode mapNode = new MapNode(new NamedNode("id", "foo"), new NamedNode("name", "The foo app"));
-        assertThat(mapNode.search("id")).isNotNull();
-        assertThat(mapNode.search("id").value()).isEqualTo("foo");
-        assertThat(mapNode.search("name").value()).isEqualTo("The foo app");
+        assertThat(mapNode.get("id")).isNotNull();
+        assertThat(mapNode.get("id").value()).isEqualTo("foo");
+        assertThat(mapNode.get("name").value()).isEqualTo("The foo app");
 
         try {
-            mapNode.search("foo");
+            mapNode.get("foo");
             Assertions.failBecauseExceptionWasNotThrown(PropertyNotFoundException.class);
         } catch (PropertyNotFoundException e) {
-            Assertions.assertThat(e.getPropertyName()).isEqualTo("foo");
+            Assertions.assertThat(e.getPropertyName()).isEqualTo("[foo]");
         }
+    }
+
+    @Test
+    public void testWithIntegerKey() {
+        MapNode mapNode = new MapNode(new NamedNode("0", "foo"));
+        assertThat(mapNode.get("0").value()).isEqualTo("foo");
     }
 }
