+++ src/test/java/org/seedstack/coffig/provider/JacksonProviderTest.java
@@ -10,7 +10,7 @@ package org.seedstack.coffig.provider;
 import org.assertj.core.api.Assertions;
 import org.junit.Before;
 import org.junit.Test;
-import org.seedstack.coffig.data.MapNode;
+import org.seedstack.coffig.MapNode;
 
 import java.io.ByteArrayInputStream;
 import java.io.InputStream;
@@ -53,7 +53,7 @@ public class JacksonProviderTest {
 
         Assertions.assertThat(mapNode).isNotNull();
         Assertions.assertThat(mapNode.value("app").value()).isEqualTo("bar");
-        Assertions.assertThat(mapNode.search("security.users.1").value()).isEqualTo("user2");
+        Assertions.assertThat(mapNode.get("security.users.1").value()).isEqualTo("user2");
     }
 
     @Test
@@ -61,6 +61,6 @@ public class JacksonProviderTest {
         MapNode mapNode = jacksonProvider.provide();
 
         Assertions.assertThat(mapNode).isNotNull();
-        Assertions.assertThat(mapNode.search("jdbc.dataSources.0.name").value()).isEqualTo("myDS");
+        Assertions.assertThat(mapNode.get("jdbc.dataSources.0.name").value()).isEqualTo("myDS");
     }
 }
