+++ src/test/java/org/seedstack/coffig/provider/SystemPropertyProviderTest.java
@@ -13,7 +13,7 @@ import mockit.integration.junit4.JMockit;
 import org.assertj.core.api.Assertions;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.seedstack.coffig.data.MapNode;
+import org.seedstack.coffig.MapNode;
 
 import java.util.Properties;
 
@@ -35,7 +35,7 @@ public class SystemPropertyProviderTest {
 
         MapNode conf = underTest.provide();
 
-        Assertions.assertThat(conf.search("system.PROFILE").value()).isEqualTo("DEV");
+        Assertions.assertThat(conf.get("system.PROFILE").value()).isEqualTo("DEV");
     }
 
     @Test
